Lift CPU Profiling
CPU Profiling naturally lives in the command execution part as well.
Move it from the underlying manager initialization up into the
rootCmd.
